requirements.txt
numpy
openai-whisper
tqdm

pydub

pandas